Move `winit::Application` to its own module
